:100: complete

+++ exercises/conversions/using_as.rs
@@ -1,3 +1,12 @@
+/*
+ * Copyright (c) 2023 by onepisYa pis1@qq.com , All Rights Reserved.
+ * @Date: 2022-12-28 21:00:00
+ * @LastEditors: onepisYa pis1@qq.com
+ * @LastEditTime: 2023-02-13 20:26:05
+ * @FilePath: /rustlings/exercises/conversions/using_as.rs
+ * 路漫漫其修远兮，吾将上下而求索。
+ * @Description:
+ */
 // Type casting in Rust is done via the usage of the `as` operator.
 // Please note that the `as` operator is not only used when type casting.
 // It also helps with renaming imports.
@@ -6,11 +15,13 @@
 // and returns the proper type.
 // Execute `rustlings hint using_as` or use the `hint` watch subcommand for a hint.
 
-// I AM NOT DONE
+// NOTE: as 操作符不仅仅用于 type  casting
+// NOTE: 也帮助用于重命名 导入
 
 fn average(values: &[f64]) -> f64 {
     let total = values.iter().sum::<f64>();
-    total / values.len()
+    // total / values.len()
+    total / values.len() as f64
 }
 
 fn main() {
